Move components to its own modules.
